Update prerequisites
Added Jupyter Notebook as a prereq. (or should this be Jupyter Lab?)

It was also unclear when I first attempted to run the pipeline whether the keys were required to be generated beforehand.  Adding a note to clarify for users that instructions to generate the keys are included in the notebook file where the key is required.