Update installation instructions in README.md
Changes made in both pip and poetry instructions:
Step 1: Added a note before running to git clone command to make certain the user is in the directory they want it to be installed into. 
Step 2. The clone command installs into ./PASS/SGMC, so I changed the cd command to reflect this.

Changes made to just the poetry instructions:
Step 5: Added a note to make certain the "sgmc_kernel" is selected for each notebook file.